python/requirements.txt
sqlalchemy
psycopg2-binary
pyyaml
python-dotenv
